hub_ai/normalization/pyproject.toml
# Poetry
[tool.poetry]
name = "normalization-module"
version = "2.1.0"
description = "An asynchronous FastAPI service using GenAI."
authors = ["Prateek K. Tiwari <Prateek_Kumar_Tiwari@external.mckinsey.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.34.3"
pydantic = "^2.7.4"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.4"
tqdm = "^4.66.4"
openai = "^1.35.3"
boto3 = "^1.34.127"
snowflake-connector-python = {extras = ["pandas"], version = "^3.10.0"}
scikit-learn = "^1.5.2"
psycopg2-binary = "^2.9.10"
tenacity = "^8.2.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Formatting
[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = ["e2e_azure", "e2e_local", "components", "utils"]
filterwarnings = ["ignore::DeprecationWarning::", "ignore::FutureWarning::"]

[tool.pydocstyle]
convention = "google"

# Testing
[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

# Omit packages in test coverage
[tool.coverage.run]
omit = ["*/azml_components/*", "*/azml_realtime_deployments/*", "*/utils/*"]

[tool.black]
line-length = 130
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''